+++ earthdaily/earthone/dynamic_compute/compute_map.py
@@ -152,6 +152,10 @@ def rasterization(*args, **kwargs):
     return None
 
 
+def morphology(*args, **kwargs):
+    return None
+
+
 class ComputeMap(dict, ABC):
     """
     A wrapper class to support operations on grafts. Proxy objects should all be
@@ -229,6 +233,7 @@ class ComputeMap(dict, ABC):
                     ("from_image_ids", from_image_ids),
                     ("convolve", convolve),
                     ("rasterization", rasterization),
+                    ("morphology", morphology),
                 ],
                 debug=True,
             )()