+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Added support for erosion and dilation.
 - Update dependencies to add support for Python 3.13.
 
 ### Changed